% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funcs.R
\name{boot_change_point}
\alias{boot_change_point}
\title{Bootstrap Estimation of Changepoint and simulated miss visits}
\usage{
boot_change_point(
  sim_data,
  sim_version = "visits",
  n_sim_trials = 100L,
  new_draw_weight = 0,
  sim_duartion_for_regression = FALSE,
  eval_criteria = "AIC",
  week_period = FALSE,
  num_cores = NULL,
  no_bootstrapping = FALSE
)
}
\arguments{
\item{sim_data}{a dataset containing the time_map, miss bins and other parameters used for the simulation.
This dataset should be created using the `prep_sim_data()` function}

\item{sim_version}{the simulation version to run. Options include "visits" (default) or "patients".}

\item{n_sim_trials}{number of trials to run in simulation of miss visits or miss patients}

\item{new_draw_weight}{a weighing parameter used to assign preference to drawing previously "missed" patients
at each time step. A value of 0 applies strict preference to drawing patients who
have been assigned to miss in prior time steps, while a value 0.5 applies equal weight
to patients who have and have not been previously selected.}

\item{num_cores}{The number of worker cores to use. If not specified will detect cores and use 1 less than the number of cores.}

\item{no_bootstrapping}{Specifies whether you want to run the simulations without bootstrapping the original dataset}
}
\description{
Bootstrap Estimation of Changepoint and simulated miss visits
}
