% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulid_timemap_function_w_ED.R
\name{build_time_map_delay}
\alias{build_time_map_delay}
\title{Builds a time map for a specific database}
\arguments{
\item{collect_tab}{A tibble with the specific setting (i.e. inpatient or outpatient), source (i.e. ccae or mdcr), and year to access.
Default is all possible combinations of setting, source, and year}

\item{db_con}{The database connection}

\item{db_path}{Path to the database}

\item{temporary}{Logical to add outpatient_keys and inpatient_keys temporarily to the database. Default is TRUE. For big databases set this to FALSE.}

\item{overwrite}{Logical to overwrite the existing outpatient_keys and inpatient_keys. Default is FALSE. For big databases set this to TRUE.}
}
\value{
A tibble with a time map containing visit level information from the database. Includes admission date, year, setting, enrolid,
standard place of care, key, discharge date, and location of care (i.e. inpatient, outpatient, or ED)
}
\description{
Builds a time map for a specific database
}
