% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsed_enrollment_function.R
\name{collapse_enrollment_delay}
\alias{collapse_enrollment_delay}
\title{Gather enrollment details information over multiple enrollment_detail tables (in parallel)}
\usage{
collapse_enrollment_delay(
  collect_tab = collect_table(),
  enrolid_list,
  vars = c(),
  db_path,
  collect_n = Inf,
  num_cores = NULL
)
}
\arguments{
\item{collect_tab}{A tibble with the specific setting (i.e. inpatient or outpatient), source (i.e. ccae or mdcr), and year to access.
Default is all possible combinations of setting, source, and year}

\item{enrolid_list}{A vector of enrolids to collect enrollment info}

\item{vars}{A character vector of variables it the core table to return. If no value is provided, enrollment will simply be collapsed into continuous enrollment periods}

\item{db_path}{Path to the database}

\item{collect_n}{The number of observations to return}

\item{num_cores}{The number of worker cores to use. If not specified will determined the number of cores based on the which ever
is the smallest value between number of rows in for collect_tab or detected number of cores - 1}
}
\value{
A tibble with all the specified variables to return over all designated enrollment_detail tables
}
\description{
Gather enrollment details information over multiple enrollment_detail tables (in parallel)
}
\examples{
test <- collapse_enrollment_delay(enrolid_list = enrolids$enrolid,
                                  db_path = db_path)

}
