% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulid_timemap_function_w_ED.R
\name{collect_table}
\alias{collect_table}
\title{Sets up a collection tibble for holding and distributing queries across a rand of tables in the database}
\usage{
collect_table(
  settings = c("inpatient", "outpatient"),
  sources = c("ccae", "mdcr"),
  years = NULL
)
}
\arguments{
\item{settings}{"inpatient", "outpatient" or c("inpatient","outpatient")}

\item{sources}{"ccae" or "mdcr" or c("ccae","mdcr")}

\item{years}{vector of years to collect. Note: if no argument is provided all years will be selected}
}
\value{
A tibble with all possible combinations of setting, source, and year
}
\description{
Sets up a collection tibble for holding and distributing queries across a rand of tables in the database
}
