% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_delay_funcs.R
\name{count_prior_events_truven}
\alias{count_prior_events_truven}
\title{Count prior visits assocatied with a given condition}
\usage{
count_prior_events_truven(time_map_data, event_name, start_day = 1L,
  by_days = 1L)
}
\arguments{
\item{time_map_data}{A timemap with indicators for specific events (i.e.
diagnoses during a given visit)}

\item{event_name}{The variable name for the even indicator}

\item{start_day}{When to start counting prior to the index}

\item{by_days}{The number of days in each bin to count by}
}
\description{
Count prior visits assocatied with a given condition
}
