% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_delay_funcs.R
\name{find_change_point}
\alias{find_change_point}
\title{Find the change point in count data}
\usage{
find_change_point(data, var_name, method = "lm", eval_criteria = "AIC",
  return_miss_only = FALSE)
}
\arguments{
\item{data}{A dataset of visit counts}

\item{var_name}{The name of the count variable to find the change-point for}

\item{method}{The method used to fit curves before and after the changepoint. Options include "lm",
"lm_quad", "lm_cube", "quad", "cube", "exp", "spline"}

\item{eval_criteria}{The evaluation criteria used to find change points}

\item{return_miss_only}{Logical argument to only return the tibbles of miss visit counts}
}
\description{
Find the change point in count data
}
