% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_functions.R
\name{find_cp_linreg}
\alias{find_cp_linreg}
\title{Find the change point in count data using linear regression models}
\usage{
find_cp_linreg(
  data,
  var_name = "n_miss_visits",
  method = "lm",
  eval_criteria = "AIC",
  return_miss_only = FALSE
)
}
\arguments{
\item{data}{A dataset of visit counts}

\item{var_name}{The name of the count variable to find the change-point for}

\item{method}{The method used to fit curves before and after the changepoint. Options include "lm",
"lm_quad", "lm_cube", "quad", "cube", "exp", "spline"}

\item{eval_criteria}{The evaluation criteria used to find change points}

\item{return_miss_only}{Logical argument to only return the tibbles of miss visit counts}
}
\description{
Find the change point in count data using linear regression models
}
\examples{
cp_result_original <- final_time_map \%>\%
count_prior_events_truven(event_name = "any_ssd", start_day = 1, by_days = 1) \%>\%
find_cp_linreg(var_name="n_miss_visits", method="lm_cube")

}
