% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpt_functions.R
\name{gather_proc_data}
\alias{gather_proc_data}
\title{Return inpatient and outpatient core information for a specific list of cpt codes over multiple core tables (in parallel)}
\usage{
gather_proc_data(
  collect_tab = collect_table(),
  cpt_codes,
  proc_vars = c("enrolid", "proc1", "svcdate", "admdate", "proc", "disdate"),
  db_path,
  collect_n = Inf,
  num_cores = NULL
)
}
\arguments{
\item{collect_tab}{A tibble with the specific setting (i.e. inpatient or outpatient), source (i.e. ccae or mdcr), and year to access.
Default is all possible combinations of setting, source, and year}

\item{cpt_codes}{The cpt codes of interest}

\item{proc_vars}{List of variables to return}

\item{db_path}{Path to the database}

\item{collect_n}{The number of observations to return}

\item{num_cores}{The number of worker cores to use. If not specified will determined the number of cores based on the which ever
is the smallest value between number of rows in for collect_tab or detected number of cores - 1}
}
\value{
A tibble with cpt code of interest for the specified list of variables
}
\description{
Return inpatient and outpatient core information for a specific list of cpt codes over multiple core tables (in parallel)
}
