% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rx_functions.R
\name{gather_rx_data}
\alias{gather_rx_data}
\title{Return information for a specific list of rx codes over multiple core tables (in parallel)}
\usage{
gather_rx_data(
  collect_tab = collect_table(),
  ndc_codes,
  rx_vars = c("enrolid", "ndcnum", "svcdate"),
  db_path,
  collect_n = Inf,
  num_cores = NULL
)
}
\arguments{
\item{collect_tab}{A tibble with the specific setting (i.e. inpatient or outpatient), source (i.e. ccae or mdcr), and year to access.
Default is all possible combinations of setting, source, and year}

\item{ndc_codes}{The rx codes of interest}

\item{rx_vars}{Path to the database}

\item{collect_n}{The number of observations to return}

\item{num_cores}{The number of worker cores to use. If not specified will determined the number of cores based on the which ever
is the smallest value between number of rows in for collect_tab or detected number of cores - 1}

\item{proc_vars}{List of variables to return}
}
\value{
A tibble with cpt code of interest for the specified list of variables
}
\description{
Return information for a specific list of rx codes over multiple core tables (in parallel)
}
