% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsed_enrollment_function.R
\name{get_collapsed_enrollment_delay}
\alias{get_collapsed_enrollment_delay}
\title{Gather enrollment details from a single enrollment_detail table from the database}
\usage{
get_collapsed_enrollment_delay(
  table_name,
  enrolid_list,
  vars = c(),
  db_path,
  collect_n = Inf
)
}
\arguments{
\item{table_name}{The specific enrollment_detail table name in the database to access}

\item{enrolid_list}{A vector of enrolids to collect enrollment info}

\item{vars}{A character vector of variables it the core table to return. If no value is provided, enrollment will simply be collapsed into continuous enrollment periods}

\item{db_path}{Path to the database}

\item{collect_n}{The number of observations to return}
}
\value{
A tibble with all the specified variables to return for one specific enrollment_detail table
}
\description{
Gather enrollment details from a single enrollment_detail table from the database
}
