% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{get_incidence_data}
\alias{get_incidence_data}
\title{Prepare a time map for incidence modeling}
\usage{
get_incidence_data(time_map, demo_data, change_point, days_before = -1)
}
\arguments{
\item{time_map}{A time map in the format of the package example}

\item{demo_data}{A dataframe of demographic data for the enrollees}

\item{change_point}{The changepoint on which to base the dataset (must be <=-1)}

\item{days_before}{Number of days before index to start classifying as a miss}
}
\value{
A dataframe containing the outcome "miss" and covariates for each row
}
\description{
Prepare a time map for incidence modeling
}
\examples{
get_incidence_data(time_map=final_time_map, demo_data=demographics, change_point=-90, days_before= -1)

}
