% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpt_functions.R
\name{gether_proc_data_outpatient}
\alias{gether_proc_data_outpatient}
\title{Return outpatient core information for a specific list of cpt codes}
\usage{
gether_proc_data_outpatient(
  table_name,
  cpt_codes,
  proc_vars = c(),
  db_path,
  collect_n = Inf
)
}
\arguments{
\item{table_name}{The specific outpatient core table to access}

\item{cpt_codes}{The cpt codes of interest}

\item{proc_vars}{List of variables to return}

\item{db_path}{Path to the database}

\item{collect_n}{The number of observations to return. Default in Inf}
}
\value{
A tibble with cpt code of interest for the specified list of variables
}
\description{
Return outpatient core information for a specific list of cpt codes
}
