% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rx_functions.R
\name{gether_rx_data}
\alias{gether_rx_data}
\title{Return rx information for a specific list of ndc codes}
\usage{
gether_rx_data(table_name, ndc_codes, rx_vars = c(), db_path, collect_n = Inf)
}
\arguments{
\item{table_name}{The specific rx core table to access}

\item{ndc_codes}{The ndc codes of interest}

\item{rx_vars}{List of variables to return}

\item{db_path}{Path to the database}

\item{collect_n}{The number of observations to return. Default in Inf}
}
\value{
A tibble with ndc code of interest for the specified list of variables
}
\description{
Return rx information for a specific list of ndc codes
}
