% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funcs.R
\name{prep_sim_data}
\alias{prep_sim_data}
\title{Prepare data for simulation}
\usage{
prep_sim_data(
  time_map_data,
  by_days = 1,
  start_day = 1,
  event_name = "any_ssd",
  cp_method = "lm_quad"
)
}
\arguments{
\item{time_map_data}{a time_map of visits}

\item{by_days}{the number of days to aggregate by in counting periods}

\item{start_day}{When to start counting prior to the index}

\item{event_name}{The variable name for the event indicator}

\item{cp_method}{The change-point method to fit visit counts}
}
\description{
Prepare data for simulation
}
