% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funcs_large_db.R
\name{run_cp_bootstrap_large_DB}
\alias{run_cp_bootstrap_large_DB}
\title{Run multiple bootstrapped change_point simulations (in parallel) for large databases}
\usage{
run_cp_bootstrap_large_DB(
  sim_data,
  sim_version = "visits",
  boot_trials = 100,
  n_sim_trials = 50,
  new_draw_weight = 0,
  num_cores = NULL,
  sim_duartion_for_regression = FALSE,
  no_bootstrapping = FALSE
)
}
\arguments{
\item{sim_data}{A dataset containing the time_map, miss bins and other parameters used for the simulation.
This dataset should be created using the `prep_sim_data()` function}

\item{sim_version}{The simulation version to run. Options include "visits" (default) or "patients"}

\item{boot_trials}{The number of bootstrapped trials to run (default is 100)}

\item{n_sim_trials}{The number of trials to run in simulation of miss visits (default is 50)}

\item{new_draw_weight}{A weighing parameter used to assign preference to drawing previously "missed" patients
at each time step. A value of 0 applies strict preference to drawing patients who
have been assigned to miss in prior time steps, while a value 0.5 applies equal weight
to patients who have and have not been previously selected}

\item{num_cores}{The number of worker cores to use. If not specified will determined the number of cores based on the which ever
is the smallest value between number of boot_trials or detected number of cores - 1}

\item{no_bootstrapping}{Specifies whether you want to run the simulations without bootstrapping the original dataset}
}
\description{
Run multiple bootstrapped change_point simulations (in parallel) for large databases
}
