% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funcs.R
\name{run_sim_miss_visits}
\alias{run_sim_miss_visits}
\title{Run multiple simulations of missed visits}
\usage{
run_sim_miss_visits(sim_data, trials = 50, sim_duartion_for_regression = FALSE)
}
\arguments{
\item{sim_data}{a dataset containing the time_map, miss bins and other parameters used for the simulation.
This dataset should be created using the `prep_sim_data()` function}

\item{trials}{number of simulationed trials to run (default is 50)}
}
\description{
Run multiple simulations of missed visits
}
