% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funcs_large_db.R
\name{run_sim_miss_visits_large_DB}
\alias{run_sim_miss_visits_large_DB}
\title{Run multiple simulations of missed visits for large databases}
\usage{
run_sim_miss_visits_large_DB(
  sim_data,
  trials = 50,
  sim_duartion_for_regression = FALSE,
  no_bootstrapping = FALSE,
  num_cores = NULL
)
}
\arguments{
\item{sim_data}{a dataset containing the time_map, miss bins and other parameters used for the simulation.
This dataset should be created using the `prep_sim_data()` function}

\item{trials}{number of simulationed trials to run (default is 50)}

\item{no_bootstrapping}{Specifies whether you want to run the simulations without bootstrapping the original dataset}

\item{num_cores}{The number of worker cores to use. If not specified will determined the number of cores based on the which ever
is the smallest value between number of trials or detected number of cores - 1}
}
\description{
Run multiple simulations of missed visits for large databases
}
