% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funcs.R
\name{sim_miss_patients}
\alias{sim_miss_patients}
\title{Simulated missed patients from an estimated set of miss bins}
\usage{
sim_miss_patients(sim_data, new_draw_weight = 0)
}
\arguments{
\item{sim_data}{a dataset containing the time_map, miss bins and other parameters used for the simulation.
This dataset should be created using the `prep_sim_data()` function}

\item{new_draw_weight}{a weighing parameter used to assign preference to drawing previously "missed" patients
at each time step. A value of 0 applies strict preference to drawing patients who
have been assigned to miss in prior time steps, while a value 0.5 applies equal weight
to patients who have and have not been previously selected}
}
\description{
Simulated missed patients from an estimated set of miss bins
}
