% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funcs_large_db.R
\name{sim_miss_visits_large_DB}
\alias{sim_miss_visits_large_DB}
\title{Simulated miss visits from an estimated set of miss bins for large databases}
\usage{
sim_miss_visits_large_DB(sim_data, sim_duartion_for_regression = FALSE)
}
\arguments{
\item{sim_data}{a dataset containing the time_map, miss bins and other parameters used for the simulation.
This dataset should be created using the `prep_sim_data()` function}
}
\description{
Simulated miss visits from an estimated set of miss bins for large databases
}
