% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_map-data.R
\docType{data}
\name{time_map}
\alias{time_map}
\title{Example time_map for running simulations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 81675 rows and 5 columns.}
\source{
created
}
\usage{
data(time_map)
}
\description{
This file contains an example of the structure of a time_map required
for running simulations. Each line of a time_map dataset contains a distinct
visit (i.e., visit to a distinct care location on a given day).
At least five different variables are required in a time_map dataset: the enrolid of the patient
(this is just a unique integer value identifying each patient), an inticator for inpatient
visits, an indicator for ed visits, an integer value denoting the days since the diagnosis
occured (this value is negative for visits prior to the index diagnosis), and an indicator
if the visit contained a particular SSD. Note: this time_map can also contain multiple different
SSD indicators. The SSD indicator to be used in the simulation will be specified in the
prep_sim_data() function.
}
\examples{
data(time_map)
}
\references{
???
}
\keyword{datasets}
