% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funcs.R
\name{prep_sim_data}
\alias{prep_sim_data}
\title{Prepare data for simulation}
\usage{
prep_sim_data(
  time_map_data,
  by_days = 1,
  start_day = 1,
  event_name = "any_ssd",
  cp_method = "lm_quad",
  specify_cp = NULL,
  eval_criteria = "AIC",
  week_period = FALSE,
  prediction_bound_for_sim = FALSE
)
}
\arguments{
\item{time_map_data}{a time_map of visits}

\item{by_days}{the number of days to aggregate by in counting periods}

\item{start_day}{When to start counting prior to the index}

\item{event_name}{The variable name for the event indicator}

\item{cp_method}{The change-point method to fit visit counts}

\item{specify_cp}{Set a specific change point you want to use instead of searching for optimal change point. Enter a postive integer value
repersenting the days before the index on which you you want to specify the change point. (e.g. 100 would be 100 days before the index)}

\item{eval_criteria}{The evaluation criteria used to find change points, if using a
linear regression method}

\item{week_period}{Logical to incorporate a "day of the week" effect into the linear model, if
method is "pettitt" of "cusum". Note this is only sensible for one-day period aggregation}

\item{prediction_bound_for_sim}{Logical to specify whether or not to use the estimated predicted value or the upper bound 90%
prediction value in the simulations. The defualt is FALSE which uses the estimated predicited value.}
}
\description{
Prepare data for simulation
}
